# docs/swagger.yaml
basePath: /
definitions:
  dto.AddToCartRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.CartItemResponse:
    properties:
      price:
        allOf:
        - $ref: '#/definitions/dto.MoneyResponse'
        description: цена за единицу
      product_id:
        type: string
      product_name:
        description: обогащённые данные из Product
        type: string
      quantity:
        type: integer
      subtotal:
        allOf:
        - $ref: '#/definitions/dto.MoneyResponse'
        description: Quantity × Price
    type: object
  dto.CartResponse:
    properties:
      created_at:
        description: ISO8601
        type: string
      id:
        type: string
      items:
        description: slice, не map (JSON-friendly)
        items:
          $ref: '#/definitions/dto.CartItemResponse'
        type: array
      items_count:
        description: количество уникальных позиций
        type: integer
      total_price:
        allOf:
        - $ref: '#/definitions/dto.MoneyResponse'
        description: структурированный Money
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      surname:
        maxLength: 255
        type: string
    required:
    - email
    - name
    type: object
  dto.MoneyResponse:
    properties:
      amount:
        description: в минимальных единицах (копейки)
        type: integer
      currency:
        description: '"RUB", "USD"'
        type: string
      formatted:
        description: '"1 500,00 ₽" — для отображения'
        type: string
    type: object
  dto.OrderItemResponse:
    properties:
      price:
        $ref: '#/definitions/dto.MoneyResponse'
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      subtotal:
        $ref: '#/definitions/dto.MoneyResponse'
    type: object
  dto.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.OrderItemResponse'
        type: array
      paid_at:
        type: string
      status:
        type: string
      total_amount:
        $ref: '#/definitions/dto.MoneyResponse'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.UpdateCartItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        description: 0 = удалить
        minimum: 0
        type: integer
    required:
    - product_id
    type: object
  dto.UpdateStatusRequest:
    properties:
      status:
        type: string
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      surname:
        maxLength: 255
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        description: ISO8601 формат
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      surname:
        type: string
      updated_at:
        type: string
    type: object
  http.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
  description: |-
    E-commerce платформа с real-time возможностями.

    ## Особенности
    - RESTful API с версионированием (/api/v1)
    - Soft delete для всех entities
    - Partial update через PATCH
    - Structured error responses
  title: Real Time System API
  version: "1.0"
paths:
  /api/v1/cart:
    delete:
      consumes:
      - application/json
      description: Очищает корзину
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Clear cart
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: Получает корзину с товарами
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get cart with items
      tags:
      - Cart
  /api/v1/cart/items:
    patch:
      consumes:
      - application/json
      description: Обновляет количество товара в корзине
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Update to cart
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Update quantity to cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Добавляет товар в корзину пользователя
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Item to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Add item to cart
      tags:
      - Cart
  /api/v1/cart/items/{product_id}:
    delete:
      consumes:
      - application/json
      description: Удаляет продукт из корзины
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Remove product from cart
      tags:
      - Cart
  /api/v1/orders:
    get:
      description: Возвращает все заказы текущего пользователя.
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Orders list
          schema:
            items:
              $ref: '#/definitions/dto.OrderResponse'
            type: array
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get user orders
      tags:
      - Orders
    post:
      description: Создаёт заказ из товаров в корзине. Уменьшает stock, очищает корзину.
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Order created
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Invalid request or empty cart
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User or product not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Insufficient stock
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Place order from cart
      tags:
      - Orders
  /api/v1/orders/{id}:
    get:
      description: Возвращает заказ с товарами. Проверяет ownership (только свои заказы).
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Order ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order found
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Invalid order ID format
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get order by ID
      tags:
      - Orders
  /api/v1/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Обновляет статус заказа. Проверяет допустимость перехода (state
        machine).
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Order ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Status updated
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Invalid request or transition
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Update order status
      tags:
      - Orders
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя. Email должен быть уникальным.
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create a new user
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Soft delete — помечает пользователя как удалённого. Данные сохраняются.
      parameters:
      - description: User ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: User deleted
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Delete user
      tags:
      - Users
    get:
      description: Возвращает пользователя по UUID.
      parameters:
      - description: User ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Частичное обновление пользователя. Передайте только поля для изменения.
      parameters:
      - description: User ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Update user (partial)
      tags:
      - Users
schemes:
- http
- https
swagger: "2.0"